Improve resolution of non-significant cells in heatmap
Display the 'float' p-value at full resolution in the noise output, and catch
any log(0) errors that arise (comparing to 0 seems to be incorrectly True for
small e-values, e.g. exponential notation, even if log can handle it).
Also tweak the color cutoffs for shades of blue in non-significant cells.